big commit, see description
Has not implemented: Return to original position
Current movement is still basic: Using lerp on position

Enemy will chase, and when the player is out of chase radius, it will go to the last known position. It will pause, then go in a random direction (search area), then pause again, and so on, 3 times. Then, it will return back to the original position. If the player reenters the chase radius then the enemy will start chasing again.